fix: add selection pills to user message
