+++ nx2/blocks/chat-ao/chat-ao.js
@@ -23,6 +23,7 @@ import { openPopoverAbove } from '../shared/chat/positioning.js';
 import { buildAttachmentItems } from '../shared/chat/files.js';
 import { renderMarkdown } from './utils/markdown.js';
 import { renderCopyButton } from '../shared/chat/copy-button.js';
+import { renderSelectionPills } from '../shared/chat/selection-pills.js';
 import '../shared/pills/pills.js';
 import '../shared/menu/menu.js';
 import '../shared/popover/popover.js';
@@ -269,14 +270,17 @@ export default class NxChatAo extends LitElement {
               @nx-show-prompts=${this._openPrompts}
             ></nx-new-chat>`
         : nothing}
-          ${this.messages?.map((msg) => html`
-            <div class="message message-${msg.role}">
-              ${msg.role === 'assistant' ? html`
-                <div class="message-content">${renderMarkdown(msg.content)}</div>
-                ${renderCopyButton(msg.content, { streaming: msg.streaming })}
-              ` : msg.content}
+          ${this.messages?.map((msg) => (msg.role === 'assistant' ? html`
+            <div class="message message-assistant">
+              <div class="message-content">${renderMarkdown(msg.content)}</div>
+              ${renderCopyButton(msg.content, { streaming: msg.streaming })}
             </div>
-          `)}
+          ` : html`
+            <div class="message message-user">
+              ${renderSelectionPills(msg)}
+              <div class="message-content">${msg.content}</div>
+            </div>
+          `))}
           ${this.thinking && !this.messages?.at(-1)?.streaming
         ? html`<div class="chat-thinking">Thinking...</div>` : nothing}
         </div>
